Assignment 3 Pres
Only rough draft and not finished. Assignment 2 not done. I also ran the Sobol script with N = 75.000 and integrated the results. For Assignment 1 I tried to get the colour of the scatter plots right, looks okay. I also chose The snow and the lrr modules, snow because its unecessary in our case and lrr being the opposite.